pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipeline-fixer-rl"
version = "1.0.0"
description = "An OpenEnv-compatible RL environment for pipeline repair."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "openenv-core>=0.2.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
]

[project.scripts]
server = "server.app:main"

[project.urls]
"Homepage" = "https://huggingface.co/spaces/Chinmay2005/broken_pipeline_fixer"

[tool.openenv]
name = "broken_pipeline_fixer"
version = "1.0.0"
tasks = ["easy", "medium", "hard"]
